pyproject.toml
[tool.poetry]
name = "fastapi-erp"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.68.0"
uvicorn = "^0.15.0"
asyncpg = "^0.29.0"
databases = "^0.8.0"
sqlalchemy = ">=1.4.42,<1.5.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
fastapi-utils = "^0.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
